fixes #890 - apoc.schema.assert are not deleted all indexes (#939)
* fixes #890 - apoc.schema.assert are not deleted all indexes

* Added EnterpriseGraphDatabaseFactory in order to test Enterprise features

+++ src/main/java/apoc/schema/Schemas.java
@@ -14,6 +14,7 @@ import org.neo4j.graphdb.schema.Schema;
 import org.neo4j.helpers.collection.Iterables;
 import org.neo4j.internal.kernel.api.*;
 import org.neo4j.internal.kernel.api.exceptions.schema.IndexNotFoundKernelException;
+import org.neo4j.internal.kernel.api.schema.constraints.ConstraintDescriptor;
 import org.neo4j.kernel.api.KernelTransaction;
 import org.neo4j.kernel.api.SilentTokenNameLookup;
 import org.neo4j.kernel.api.Statement;
@@ -21,6 +22,7 @@ import org.neo4j.procedure.*;
 
 import java.util.*;
 import java.util.concurrent.ExecutionException;
+import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import java.util.stream.StreamSupport;
 
@@ -35,7 +37,7 @@ public class Schemas {
 
     @Procedure(value = "apoc.schema.assert", mode = Mode.SCHEMA)
     @Description("apoc.schema.assert({indexLabel:[[indexKeys]], ...}, {constraintLabel:[constraintKeys], ...}, dropExisting : true) yield label, key, keys, unique, action - drops all other existing indexes and constraints when `dropExisting` is `true` (default is `true`), and asserts that at the end of the operation the given indexes and unique constraints are there, each label:key pair is considered one constraint/label. Non-constraint indexes can define compound indexes with label:[key1,key2...] pairings.")
-    public Stream<AssertSchemaResult> schemaAssert(@Name("indexes") Map<String, List<Object>> indexes, @Name("constraints") Map<String, List<String>> constraints, @Name(value = "dropExisting", defaultValue = "true") boolean dropExisting) throws ExecutionException, InterruptedException {
+    public Stream<AssertSchemaResult> schemaAssert(@Name("indexes") Map<String, List<Object>> indexes, @Name("constraints") Map<String, List<Object>> constraints, @Name(value = "dropExisting", defaultValue = "true") boolean dropExisting) throws ExecutionException, InterruptedException {
         return Stream.concat(
                 assertIndexes(indexes, dropExisting).stream(),
                 assertConstraints(constraints, dropExisting).stream());
@@ -71,19 +73,16 @@ public class Schemas {
         return constraintsExistsForRelationship(type, propertyNames);
     }
 
-    public List<AssertSchemaResult> assertConstraints(Map<String, List<String>> constraints0, boolean dropExisting) throws ExecutionException, InterruptedException {
-        Map<String, List<String>> constraints = copy(constraints0);
+    public List<AssertSchemaResult> assertConstraints(Map<String, List<Object>> constraints0, boolean dropExisting) throws ExecutionException, InterruptedException {
+        Map<String, List<Object>> constraints = copyMapOfObjects(constraints0);
         List<AssertSchemaResult> result = new ArrayList<>(constraints.size());
         Schema schema = db.schema();
 
         for (ConstraintDefinition definition : schema.getConstraints()) {
-            if (!definition.isConstraintType(ConstraintType.UNIQUENESS)) continue;
-
+            if (!(definition.isConstraintType(ConstraintType.UNIQUENESS) || definition.isConstraintType(ConstraintType.NODE_KEY))) continue;
             String label = definition.getLabel().name();
-            String key = Iterables.single(definition.getPropertyKeys());
-
-            AssertSchemaResult info = new AssertSchemaResult(label, key).unique();
-            if (!constraints.containsKey(label) || !constraints.get(label).remove(key)) {
+            AssertSchemaResult info = new AssertSchemaResult(label, Iterables.asList(definition.getPropertyKeys())).unique();
+            if (!constraints.containsKey(label) || !constraints.get(label).remove(info.key)) {
                 if (dropExisting) {
                     definition.drop();
                     info.dropped();
@@ -92,14 +91,30 @@ public class Schemas {
             result.add(info);
         }
 
-        for (Map.Entry<String, List<String>> constraint : constraints.entrySet()) {
-            for (String key : constraint.getValue()) {
-                schema.constraintFor(label(constraint.getKey())).assertPropertyIsUnique(key).create();
-                result.add(new AssertSchemaResult(constraint.getKey(), key).unique().created());
+        for (Map.Entry<String, List<Object>> constraint : constraints.entrySet()) {
+            for (Object key : constraint.getValue()) {
+                if (key instanceof String) {
+                    result.add(createUniqueConstraint(schema, constraint.getKey(), key.toString()));
+                } else if (key instanceof List) {
+                    result.add(createNodeKeyConstraint(constraint.getKey(), (List<Object>) key));
+                }
             }
         }
-
         return result;
+    }
+
+    private AssertSchemaResult createNodeKeyConstraint(String lbl, List<Object> keys) {
+        String keyProperties = keys.stream()
+                .map( property -> String.format("n.`%s`", property))
+                .collect( Collectors.joining( "," ) );
+        db.execute(String.format("CREATE CONSTRAINT ON (n:`%s`) ASSERT (%s) IS NODE KEY", lbl, keyProperties)).close();
+        List<String> keysToSting = keys.stream().map(Object::toString).collect(Collectors.toList());
+        return new AssertSchemaResult(lbl, keysToSting).unique().created();
+    }
+
+    private AssertSchemaResult createUniqueConstraint(Schema schema, String lbl, String key) {
+        schema.constraintFor(label(lbl)).assertPropertyIsUnique(key).create();
+        return new AssertSchemaResult(lbl, key).unique().created();
     }
 
     public List<AssertSchemaResult> assertIndexes(Map<String, List<Object>> indexes0, boolean dropExisting) throws ExecutionException, InterruptedException, IllegalArgumentException {
@@ -255,17 +270,25 @@ public class Schemas {
      *
      * @return
      */
-    private Stream<IndexConstraintNodeInfo> indexesAndConstraintsForNode() throws IndexNotFoundKernelException {
-        Schema schema = db.schema();
+    private Stream<IndexConstraintNodeInfo> indexesAndConstraintsForNode() {
         try ( Statement ignore = tx.acquireStatement() ) {
             TokenRead tokenRead = tx.tokenRead();
             TokenNameLookup tokens = new SilentTokenNameLookup(tokenRead);
 
             SchemaRead schemaRead = tx.schemaRead();
             Iterable<IndexReference> indexesIterator = () -> schemaRead.indexesGetAll();
-            return StreamSupport.stream(indexesIterator.spliterator(), false)
-                   .map(indexReference -> this.nodeInfoFromIndexDefinition(indexReference, schemaRead, tokens))
+
+            Iterable<ConstraintDescriptor> constraintsIterator = () -> schemaRead.constraintsGetAll();
+            Stream<IndexConstraintNodeInfo> constraintNodeInfoStream = StreamSupport.stream(constraintsIterator.spliterator(), false)
+                    .filter(constraintDescriptor -> constraintDescriptor.type().equals(ConstraintDescriptor.Type.EXISTS))
+                    .map(constraintDescriptor -> this.nodeInfoFromConstraintDescriptor(constraintDescriptor, tokens))
                     .sorted(Comparator.comparing(i -> i.label));
+
+            Stream<IndexConstraintNodeInfo> indexNodeInfoStream = StreamSupport.stream(indexesIterator.spliterator(), false)
+                    .map(indexReference -> this.nodeInfoFromIndexDefinition(indexReference, schemaRead, tokens))
+                    .sorted(Comparator.comparing(i -> i.label));
+
+            return Stream.of(constraintNodeInfoStream, indexNodeInfoStream).flatMap(e -> e);
         }
     }
 
@@ -282,11 +305,39 @@ public class Schemas {
                 .map(this::relationshipInfoFromConstraintDefinition);
     }
 
+
+    /**
+     * ConstraintInfo info from ConstraintDescriptor
+     *
+     * @param constraintDescriptor
+     * @param tokens
+     * @return
+     */
+    private IndexConstraintNodeInfo nodeInfoFromConstraintDescriptor(ConstraintDescriptor constraintDescriptor, TokenNameLookup tokens) {
+        String labelName =  tokens.labelGetName(constraintDescriptor.schema().keyId());
+        List<String> properties = new ArrayList<>();
+        Arrays.stream(constraintDescriptor.schema().getPropertyIds()).forEach((i) -> properties.add(tokens.propertyKeyGetName(i)));
+        return new IndexConstraintNodeInfo(
+                // Pretty print for index name
+                String.format(":%s(%s)", labelName, StringUtils.join(properties, ",")),
+                labelName,
+                properties,
+                StringUtils.EMPTY,
+                ConstraintType.NODE_PROPERTY_EXISTENCE.toString(),
+                "NO FAILURE",
+                0,
+                0,
+                0,
+                constraintDescriptor.userDescription(tokens)
+        );
+    }
+
     /**
      * Index info from IndexDefinition
      *
      * @param indexReference
      * @param schemaRead
+     * @param tokens
      * @return
      */
     private IndexConstraintNodeInfo nodeInfoFromIndexDefinition(IndexReference indexReference, SchemaRead schemaRead, TokenNameLookup tokens){
